requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cloudinary==1.44.0
colorama==0.4.6
cryptography==45.0.3
distro==1.9.0
dj-database-url==3.0.0
Django==5.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
django-leaflet==0.32.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-webpush==0.3.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
Faker==37.3.0
frozenlist==1.6.2
gunicorn==23.0.0
h11==0.16.0
http_ece==1.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
multidict==6.4.4
openai==1.84.0
packaging==25.0
phonenumbers==9.0.10
propcache==0.3.1
psycopg2-binary>=2.9
py-vapid==1.9.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python-dotenv==1.1.0
pywebpush==2.0.3
redis==6.2.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.2.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
yarl==1.20.0
